+++ Mesa04/Models/TipoRegistroNacional.cs
@@ -41,7 +41,7 @@ namespace Mesa04.Models
         
 
         //operação (metodo customizado) para retornar o total de vendas do departamento
-        public double TotalSales(DateTime inicial, DateTime final)
+        public decimal TotalSales(DateTime inicial, DateTime final)
         {
             return Clientes.Sum(cliente => cliente.TotalSales(inicial, final));
         }
